problem 3 solved: result 6857

+++ src/de/bjrke/euler/problem0007/Problem0007.scala
@@ -2,19 +2,19 @@ package de.bjrke.euler.problem0007
 
 object Problem0007 {
   def main(args : Array[String]) : Unit = {
-    val a: List[Int] = List( 3, 2)
+    val a: List[Long] = List( 3, 2)
     println(findPrime(10001,a))
   }  
   
-  def findPrime(number: Int, primes: List[Int]) : Int = {
+  def findPrime(number: Long, primes: List[Long]) : Long = {
     if ( primes.size == number ) {
       primes.head
     } else {
       findPrime(number, nextPrime(primes) :: primes)
     }
   }
-  def nextPrime( previous: List[Int] ) : Int = nextPrime( previous, previous.head )
-  def nextPrime( previous: List[Int], lastTested: Int ) : Int = {
+  def nextPrime( previous: List[Long] ) : Long = nextPrime( previous, previous.head )
+  def nextPrime( previous: List[Long], lastTested: Long ) : Long = {
     val cand = lastTested + 2
     if ( previous.exists ( cand % _ == 0 ) ) {
       nextPrime( previous, cand )
